+++ cmd/agent/app/flags_test.go
@@ -34,8 +34,6 @@ func TestBingFlags(t *testing.T) {
 	v.BindPFlags(command.PersistentFlags())
 
 	err := command.ParseFlags([]string{
-		"--collector.host-port=1.2.3.4:555,1.2.3.4:666",
-		"--discovery.min-peers=42",
 		"--http-server.host-port=:8080",
 		"--processor.jaeger-binary.server-host-port=:1111",
 		"--processor.jaeger-binary.server-max-packet-size=4242",
@@ -46,8 +44,6 @@ func TestBingFlags(t *testing.T) {
 
 	b.InitFromViper(v)
 	assert.Equal(t, 3, len(b.Processors))
-	assert.Equal(t, []string{"1.2.3.4:555", "1.2.3.4:666"}, b.CollectorHostPorts)
-	assert.Equal(t, 42, b.DiscoveryMinPeers)
 	assert.Equal(t, ":8080", b.HTTPServer.HostPort)
 	assert.Equal(t, ":1111", b.Processors[2].Server.HostPort)
 	assert.Equal(t, 4242, b.Processors[2].Server.MaxPacketSize)