+++ cmd/all-in-one/main.go
@@ -36,12 +36,18 @@ import (
 	"github.com/uber/tchannel-go"
 	"github.com/uber/tchannel-go/thrift"
 	"go.uber.org/zap"
+	"google.golang.org/grpc"
 
 	agentApp "github.com/jaegertracing/jaeger/cmd/agent/app"
+	agentRep "github.com/jaegertracing/jaeger/cmd/agent/app/reporter"
+	agentGrpcRep "github.com/jaegertracing/jaeger/cmd/agent/app/reporter/grpc"
+	agentTchanRep "github.com/jaegertracing/jaeger/cmd/agent/app/reporter/tchannel"
 	basic "github.com/jaegertracing/jaeger/cmd/builder"
 	collectorApp "github.com/jaegertracing/jaeger/cmd/collector/app"
 	collector "github.com/jaegertracing/jaeger/cmd/collector/app/builder"
+	"github.com/jaegertracing/jaeger/cmd/collector/app/grpcserver"
 	"github.com/jaegertracing/jaeger/cmd/collector/app/sampling"
+	"github.com/jaegertracing/jaeger/cmd/collector/app/sampling/strategystore"
 	"github.com/jaegertracing/jaeger/cmd/collector/app/zipkin"
 	"github.com/jaegertracing/jaeger/cmd/env"
 	"github.com/jaegertracing/jaeger/cmd/flags"
@@ -121,14 +127,19 @@ func main() {
 			if err != nil {
 				logger.Fatal("Failed to create dependency reader", zap.Error(err))
 			}
-			samplingHandler := initializeSamplingHandler(strategyStoreFactory, v, metricsFactory, logger)
+
+			strategyStoreFactory.InitFromViper(v)
+			strategyStore := initSamplingStrategyStore(strategyStoreFactory, metricsFactory, logger)
 
 			aOpts := new(agentApp.Builder).InitFromViper(v)
+			repOpts := new(agentRep.Options).InitFromViper(v)
+			tchannelRepOpts := agentTchanRep.NewBuilder().InitFromViper(v, logger)
+			grpcRepOpts := new(agentGrpcRep.Options).InitFromViper(v)
 			cOpts := new(collector.CollectorOptions).InitFromViper(v)
 			qOpts := new(queryApp.QueryOptions).InitFromViper(v)
 
-			startAgent(aOpts, cOpts, logger, metricsFactory)
-			startCollector(cOpts, spanWriter, logger, metricsFactory, samplingHandler, hc)
+			startAgent(aOpts, repOpts, tchannelRepOpts, grpcRepOpts, cOpts, logger, metricsFactory)
+			startCollector(cOpts, spanWriter, logger, metricsFactory, strategyStore, hc)
 			startQuery(qOpts, spanReader, dependencyReader, logger, metricsFactory, mBldr, hc)
 			hc.Ready()
 			<-signalsChannel
@@ -156,6 +167,9 @@ func main() {
 		flags.AddFlags,
 		storageFactory.AddFlags,
 		agentApp.AddFlags,
+		agentRep.AddFlags,
+		agentTchanRep.AddFlags,
+		agentGrpcRep.AddFlags,
 		collector.AddFlags,
 		queryApp.AddFlags,
 		pMetrics.AddFlags,
@@ -170,16 +184,21 @@ func main() {
 
 func startAgent(
 	b *agentApp.Builder,
+	repOpts *agentRep.Options,
+	tchanRep *agentTchanRep.Builder,
+	grpcRepOpts *agentGrpcRep.Options,
 	cOpts *collector.CollectorOptions,
 	logger *zap.Logger,
 	baseFactory metrics.Factory,
 ) {
 	metricsFactory := baseFactory.Namespace("agent", nil)
 
-	if len(b.CollectorHostPorts) == 0 {
-		b.CollectorHostPorts = append(b.CollectorHostPorts, fmt.Sprintf("127.0.0.1:%d", cOpts.CollectorPort))
+	cp, err := createCollectorProxy(cOpts, repOpts, tchanRep, grpcRepOpts, logger, metricsFactory)
+	if err != nil {
+		logger.Fatal("Could not create collector proxy", zap.Error(err))
 	}
-	agent, err := b.WithMetricsFactory(metricsFactory).CreateAgent(logger)
+
+	agent, err := b.CreateAgent(cp, logger, baseFactory)
 	if err != nil {
 		logger.Fatal("Unable to initialize Jaeger Agent", zap.Error(err))
 	}
@@ -190,12 +209,32 @@ func startAgent(
 	}
 }
 
+func createCollectorProxy(
+	cOpts *collector.CollectorOptions,
+	repOpts *agentRep.Options,
+	tchanRepOpts *agentTchanRep.Builder,
+	grpcRepOpts *agentGrpcRep.Options,
+	logger *zap.Logger,
+	mFactory metrics.Factory,
+) (agentApp.CollectorProxy, error) {
+	switch repOpts.ReporterType {
+	case agentRep.GRPC:
+		grpcRepOpts.CollectorHostPort = append(grpcRepOpts.CollectorHostPort, fmt.Sprintf("127.0.0.1:%d", cOpts.CollectorGRPCPort))
+		return agentGrpcRep.NewCollectorProxy(grpcRepOpts, logger)
+	case agentRep.TCHANNEL:
+		tchanRepOpts.CollectorHostPorts = append(tchanRepOpts.CollectorHostPorts, fmt.Sprintf("127.0.0.1:%d", cOpts.CollectorPort))
+		return agentTchanRep.NewCollectorProxy(tchanRepOpts, mFactory, logger)
+	default:
+		return nil, errors.New(fmt.Sprintf("unknown reporter type %s", string(repOpts.ReporterType)))
+	}
+}
+
 func startCollector(
 	cOpts *collector.CollectorOptions,
 	spanWriter spanstore.Writer,
 	logger *zap.Logger,
 	baseFactory metrics.Factory,
-	samplingHandler sampling.Handler,
+	strategyStore strategystore.StrategyStore,
 	hc *healthcheck.HealthCheck,
 ) {
 	metricsFactory := baseFactory.Namespace("collector", nil)
@@ -209,38 +248,51 @@ func startCollector(
 	if err != nil {
 		logger.Fatal("Unable to set up builder", zap.Error(err))
 	}
-	ch, err := tchannel.NewChannel("jaeger-collector", &tchannel.ChannelOptions{})
-	if err != nil {
-		logger.Fatal("Unable to create new TChannel", zap.Error(err))
-	}
-	server := thrift.NewServer(ch)
-	zipkinSpansHandler, jaegerBatchesHandler := spanBuilder.BuildHandlers()
-	server.Register(jc.NewTChanCollectorServer(jaegerBatchesHandler))
-	server.Register(zc.NewTChanZipkinCollectorServer(zipkinSpansHandler))
-	server.Register(sc.NewTChanSamplingManagerServer(samplingHandler))
-	portStr := ":" + strconv.Itoa(cOpts.CollectorPort)
-	listener, err := net.Listen("tcp", portStr)
-	if err != nil {
-		logger.Fatal("Unable to start listening on channel", zap.Error(err))
-	}
-	ch.Serve(listener)
-	logger.Info("Starting jaeger-collector TChannel server", zap.Int("port", cOpts.CollectorPort))
 
-	r := mux.NewRouter()
-	apiHandler := collectorApp.NewAPIHandler(jaegerBatchesHandler)
-	apiHandler.RegisterRoutes(r)
-	httpPortStr := ":" + strconv.Itoa(cOpts.CollectorHTTPPort)
-	recoveryHandler := recoveryhandler.NewRecoveryHandler(logger, true)
+	zipkinSpansHandler, jaegerBatchesHandler, grpcHandler := spanBuilder.BuildHandlers()
 
-	go startZipkinHTTPAPI(logger, cOpts.CollectorZipkinHTTPPort, zipkinSpansHandler, recoveryHandler)
-
-	logger.Info("Starting jaeger-collector HTTP server", zap.Int("http-port", cOpts.CollectorHTTPPort))
-	go func() {
-		if err := http.ListenAndServe(httpPortStr, recoveryHandler(r)); err != nil {
-			logger.Fatal("Could not launch jaeger-collector HTTP server", zap.Error(err))
+	{
+		ch, err := tchannel.NewChannel("jaeger-collector", &tchannel.ChannelOptions{})
+		if err != nil {
+			logger.Fatal("Unable to create new TChannel", zap.Error(err))
 		}
-		hc.Set(healthcheck.Unavailable)
-	}()
+		server := thrift.NewServer(ch)
+		server.Register(jc.NewTChanCollectorServer(jaegerBatchesHandler))
+		server.Register(zc.NewTChanZipkinCollectorServer(zipkinSpansHandler))
+		server.Register(sc.NewTChanSamplingManagerServer(sampling.NewHandler(strategyStore)))
+		portStr := ":" + strconv.Itoa(cOpts.CollectorPort)
+		listener, err := net.Listen("tcp", portStr)
+		if err != nil {
+			logger.Fatal("Unable to start listening on channel", zap.Error(err))
+		}
+		logger.Info("Starting jaeger-collector TChannel server", zap.Int("port", cOpts.CollectorPort))
+		ch.Serve(listener)
+	}
+
+	{
+		grpcserver.StartGRPCCollector(cOpts.CollectorGRPCPort, grpc.NewServer(), grpcHandler, strategyStore, logger,
+			func(err error) {
+				logger.Fatal("gRPC collector failed", zap.Error(err))
+			})
+	}
+
+	{
+		r := mux.NewRouter()
+		apiHandler := collectorApp.NewAPIHandler(jaegerBatchesHandler)
+		apiHandler.RegisterRoutes(r)
+		httpPortStr := ":" + strconv.Itoa(cOpts.CollectorHTTPPort)
+		recoveryHandler := recoveryhandler.NewRecoveryHandler(logger, true)
+
+		go startZipkinHTTPAPI(logger, cOpts.CollectorZipkinHTTPPort, zipkinSpansHandler, recoveryHandler)
+
+		logger.Info("Starting jaeger-collector HTTP server", zap.Int("http-port", cOpts.CollectorHTTPPort))
+		go func() {
+			if err := http.ListenAndServe(httpPortStr, recoveryHandler(r)); err != nil {
+				logger.Fatal("Could not launch jaeger-collector HTTP server", zap.Error(err))
+			}
+			hc.Set(healthcheck.Unavailable)
+		}()
+	}
 }
 
 func startZipkinHTTPAPI(
@@ -319,13 +371,11 @@ func startQuery(
 	}()
 }
 
-func initializeSamplingHandler(
+func initSamplingStrategyStore(
 	samplingStrategyStoreFactory *ss.Factory,
-	v *viper.Viper,
 	metricsFactory metrics.Factory,
 	logger *zap.Logger,
-) sampling.Handler {
-	samplingStrategyStoreFactory.InitFromViper(v)
+) strategystore.StrategyStore {
 	if err := samplingStrategyStoreFactory.Initialize(metricsFactory, logger); err != nil {
 		logger.Fatal("Failed to init sampling strategy store factory", zap.Error(err))
 	}
@@ -333,5 +383,5 @@ func initializeSamplingHandler(
 	if err != nil {
 		logger.Fatal("Failed to create sampling strategy store", zap.Error(err))
 	}
-	return sampling.NewHandler(strategyStore)
+	return strategyStore
 }