+++ cmd/ingester/app/consumer/consumer.go
@@ -29,10 +29,11 @@ import (
 
 // Params are the parameters of a Consumer
 type Params struct {
-	ProcessorFactory ProcessorFactory
-	Factory          metrics.Factory
-	Logger           *zap.Logger
-	InternalConsumer consumer.Consumer
+	ProcessorFactory      ProcessorFactory
+	MetricsFactory        metrics.Factory
+	Logger                *zap.Logger
+	InternalConsumer      consumer.Consumer
+	DeadlockCheckInterval time.Duration
 }
 
 // Consumer uses sarama to consume and handle messages from kafka
@@ -46,6 +47,7 @@ type Consumer struct {
 	deadlockDetector deadlockDetector
 
 	partitionIDToState map[int32]*consumerState
+	partitionsHeld     metrics.Counter
 }
 
 type consumerState struct {
@@ -55,14 +57,15 @@ type consumerState struct {
 
 // New is a constructor for a Consumer
 func New(params Params) (*Consumer, error) {
-	deadlockDetector := newDeadlockDetector(params.Factory, params.Logger, time.Minute)
+	deadlockDetector := newDeadlockDetector(params.MetricsFactory, params.Logger, params.DeadlockCheckInterval)
 	return &Consumer{
-		metricsFactory:     params.Factory,
+		metricsFactory:     params.MetricsFactory,
 		logger:             params.Logger,
 		internalConsumer:   params.InternalConsumer,
 		processorFactory:   params.ProcessorFactory,
 		deadlockDetector:   deadlockDetector,
 		partitionIDToState: make(map[int32]*consumerState),
+		partitionsHeld:     partitionsHeld(params.MetricsFactory),
 	}, nil
 }
 
@@ -100,6 +103,8 @@ func (c *Consumer) Close() error {
 
 func (c *Consumer) handleMessages(pc sc.PartitionConsumer) {
 	c.logger.Info("Starting message handler", zap.Int32("partition", pc.Partition()))
+	c.partitionsHeld.Inc(1)
+	defer c.partitionsHeld.Inc(-1)
 	c.partitionIDToState[pc.Partition()].wg.Add(1)
 	defer c.partitionIDToState[pc.Partition()].wg.Done()
 	defer c.closePartition(pc)