+++ examples/hotrod/pkg/tracing/init.go
@@ -16,14 +16,11 @@ package tracing
 
 import (
 	"fmt"
-	"strings"
 	"time"
 
 	"github.com/opentracing/opentracing-go"
-	"github.com/uber/jaeger-client-go"
 	"github.com/uber/jaeger-client-go/config"
 	"github.com/uber/jaeger-client-go/rpcmetrics"
-	"github.com/uber/jaeger-client-go/transport"
 	"github.com/uber/jaeger-lib/metrics"
 	"go.uber.org/zap"
 
@@ -31,36 +28,21 @@ import (
 )
 
 // Init creates a new instance of Jaeger tracer.
-func Init(serviceName string, metricsFactory metrics.Factory, logger log.Factory, backendHostPort string) opentracing.Tracer {
-	cfg := config.Configuration{
-		Sampler: &config.SamplerConfig{
-			Type:  "const",
-			Param: 1,
-		},
+func Init(serviceName string, metricsFactory metrics.Factory, logger log.Factory) opentracing.Tracer {
+	cfg, err := config.FromEnv()
+	if err != nil {
+		logger.Bg().Fatal("cannot parse Jaeger env vars", zap.Error(err))
 	}
+	cfg.ServiceName = serviceName
+	cfg.Sampler.Type = "const"
+	cfg.Sampler.Param = 1
+
 	// TODO(ys) a quick hack to ensure random generators get different seeds, which are based on current time.
 	time.Sleep(100 * time.Millisecond)
 	jaegerLogger := jaegerLoggerAdapter{logger.Bg()}
-	var sender jaeger.Transport
-	if strings.HasPrefix(backendHostPort, "http://") {
-		sender = transport.NewHTTPTransport(
-			backendHostPort,
-			transport.HTTPBatchSize(1),
-		)
-	} else {
-		if s, err := jaeger.NewUDPTransport(backendHostPort, 0); err != nil {
-			logger.Bg().Fatal("cannot initialize UDP sender", zap.Error(err))
-		} else {
-			sender = s
-		}
-	}
-	tracer, _, err := cfg.New(
-		serviceName,
-		config.Reporter(jaeger.NewRemoteReporter(
-			sender,
-			jaeger.ReporterOptions.BufferFlushInterval(1*time.Second),
-			jaeger.ReporterOptions.Logger(jaegerLogger),
-		)),
+
+	metricsFactory = metricsFactory.Namespace(serviceName, nil)
+	tracer, _, err := cfg.NewTracer(
 		config.Logger(jaegerLogger),
 		config.Metrics(metricsFactory),
 		config.Observer(rpcmetrics.NewObserver(metricsFactory, rpcmetrics.DefaultNameNormalizer)),