+++ examples/hotrod/services/customer/server.go
@@ -36,13 +36,13 @@ type Server struct {
 }
 
 // NewServer creates a new customer.Server
-func NewServer(hostPort string, tracer opentracing.Tracer, metricsFactory metrics.Factory, logger log.Factory, jAgentHostPort string) *Server {
+func NewServer(hostPort string, tracer opentracing.Tracer, metricsFactory metrics.Factory, logger log.Factory) *Server {
 	return &Server{
 		hostPort: hostPort,
 		tracer:   tracer,
 		logger:   logger,
 		database: newDatabase(
-			tracing.Init("mysql", metricsFactory.Namespace("mysql", nil), logger, jAgentHostPort),
+			tracing.Init("mysql", metricsFactory, logger),
 			logger.With(zap.String("component", "mysql")),
 		),
 	}