+++ CHANGELOG.md
@@ -1,12 +1,56 @@
 Changes by Version
 ==================
 
-1.8.0 (unreleased)
+1.9.0 (unreleased)
 ------------------
 
 #### Backend Changes
 
 ##### Breaking Changes
+
+- Rename tcollector proxy metric in agent ([#1182](https://github.com/jaegertracing/jaeger/pull/1182), [@pavolloffay](https://github.com/pavolloffay))
+
+The following metric:
+```
+jaeger_http_server_errors{source="tcollector-proxy",status="5xx"}
+```
+has been renamed to:
+```
+jaeger_http_server_errors{source="collector-proxy",status="5xx"}
+```
+
+##### New Features
+
+##### Bug fixes, Minor Improvements
+
+#### UI Changes
+
+##### New Features
+
+##### Bug Fixes, Minor Improvements
+
+
+1.8.0 (2018-11-12)
+------------------
+
+#### Backend Changes
+
+##### Breaking Changes
+
+- Refactor agent configuration ([#1092](https://github.com/jaegertracing/jaeger/pull/1092), [@pavolloffay](https://github.com/pavolloffay))
+
+The following agent flags has has been deprecated in order to support multiple reporters:
+```bash
+--collector.host-port
+--discovery.conn-check-timeout
+--discovery.min-peers
+```
+New flags:
+```bash
+--reporter.tchannel.host-port
+--reporter.tchannel.discovery.conn-check-timeout
+--reporter.tchannel.discovery.min-peers
+```
 
 - Various changes around metrics produced by jaeger-query: Names scoped to the query component, generated for all span readers (not just ES), consolidate query metrics and include result tag ([#1074](https://github.com/jaegertracing/jaeger/pull/1074), [#1075](https://github.com/jaegertracing/jaeger/pull/1075) and [#1096](https://github.com/jaegertracing/jaeger/pull/1096), [@objectiser](https://github.com/objectiser))
 
@@ -31,6 +75,39 @@ jaeger_query_requests{operation="find_traces",result="ok"} 2
 jaeger_query_responses_bucket{operation="find_traces",le="0.005"} 2
 ```
 
+##### New Features
+
+- Configurable deadlock detector interval for ingester ([#1134](https://github.com/jaegertracing/jaeger/pull/1134), [@marqc](https://github.com/marqc))
+- Emit spans for elastic storage backend ([#1128](https://github.com/jaegertracing/jaeger/pull/1128), [@annanay25](https://github.com/annanay25))
+- Allow to use TLS certificates for Elasticsearch authentication ([#1139](https://github.com/jaegertracing/jaeger/pull/1139), [@clyang82](https://github.com/clyang82))
+- Add ingester metrics, healthcheck and rename Kafka cli flags ([#1094](https://github.com/jaegertracing/jaeger/pull/1094), [@ledor473](https://github.com/ledor473))
+- Add a metric for number of partitions held ([#1154](https://github.com/jaegertracing/jaeger/pull/1154), [@vprithvi](https://github.com/vprithvi))
+- Log jaeger-collector tchannel port ([#1136](https://github.com/jaegertracing/jaeger/pull/1136), [@mindaugasrukas](https://github.com/mindaugasrukas))
+- Support tracer env based initialization in hotrod ([#1115](https://github.com/jaegertracing/jaeger/pull/1115), [@eundoosong](https://github.com/eundoosong))
+- Publish ingester as binaries and docker image ([#1086](https://github.com/jaegertracing/jaeger/pull/1086), [@ledor473](https://github.com/ledor473))
+- Use Go 1.11 ([#1104](https://github.com/jaegertracing/jaeger/pull/1104), [@isaachier](https://github.com/isaachier))
+- Tag images with commit SHA and publish to `-snapshot` repository ([#1082](https://github.com/jaegertracing/jaeger/pull/1082), [@pavolloffay](https://github.com/pavolloffay))
+
+##### Bug fixes, Minor Improvements
+
+- Fix child span context while tracing cassandra queries ([#1131](https://github.com/jaegertracing/jaeger/pull/1131), [@annanay25](https://github.com/annanay25))
+- Deadlock detector hack for Kafka driver instability ([#1087](https://github.com/jaegertracing/jaeger/pull/1087), [@vprithvi](https://github.com/vprithvi))
+- Fix processor overriding data in a buffer ([#1099](https://github.com/jaegertracing/jaeger/pull/1099), [@pavolloffay](https://github.com/pavolloffay))
+
+#### UI Changes
+
+##### New Features
+
+- Span Search - Highlight search results ([#238](https://github.com/jaegertracing/jaeger-ui/pull/238)), [@davit-y](https://github.com/davit-y)
+- Span Search - Improve search logic ([#237](https://github.com/jaegertracing/jaeger-ui/pull/237)),  [@davit-y](https://github.com/davit-y)
+- Span Search - Add result count, navigation and clear buttons ([#234](https://github.com/jaegertracing/jaeger-ui/pull/234)), [@davit-y](https://github.com/davit-y)
+
+##### Bug Fixes, Minor Improvements
+
+- Use correct duration format for scatter plot ([#266](https://github.com/jaegertracing/jaeger-ui/pull/266)), [@tiffon](https://github.com/tiffon))
+- Fix collapse all issues ([#264](https://github.com/jaegertracing/jaeger-ui/pull/264)), [@tiffon](https://github.com/tiffon))
+- Use a moderately sized canvas for the span graph ([#257](https://github.com/jaegertracing/jaeger-ui/pull/257)), [@tiffon](https://github.com/tiffon))
+
 
 1.7.0 (2018-09-19)
 ------------------
@@ -45,7 +122,7 @@ jaeger_query_responses_bucket{operation="find_traces",le="0.005"} 2
 
 #### Backend Changes
 
-##### Breaking changes
+##### Breaking Changes
 
 - `jaeger-standalone` binary has been renamed to `jaeger-all-in-one`. This change also includes package rename from `standalone` to `all-in-one` ([#1062](https://github.com/jaegertracing/jaeger/pull/1062), [@pavolloffay](https://github.com/pavolloffay))
 
@@ -64,7 +141,7 @@ jaeger_query_responses_bucket{operation="find_traces",le="0.005"} 2
 
 #### Backend Changes
 
-##### Breaking Changes!!!
+##### Breaking Changes
 
 - The storage implementations no longer write the parentSpanID field to storage (#856).
   If you are upgrading to this version, **you must upgrade query service first**!
@@ -164,7 +241,7 @@ jaeger_query_responses_bucket{operation="find_traces",le="0.005"} 2
 
 - Support traces with >10k spans in Elasticsearch (#668) - thanks to @sramakr
 
-##### Fixes
+##### Bug Fixes, Minor Improvements
 
 - Allow slash '/' in service names (#586)
 - Log errors from HotROD services (#769)
@@ -186,7 +263,7 @@ jaeger_query_responses_bucket{operation="find_traces",le="0.005"} 2
 - Collect metrics from ES bulk service (#688) <Pavol Loffay>
 - Allow zero replicas for Elasticsearch (#754) <bharat-p>
 
-##### Fixes
+##### Bug Fixes, Minor Improvements
 
 - Apply namespace when creating Prometheus metrics factory (fix for #732) (#733) <Yuri Shkuro>
 - Disable double compression on Prom Handler - fixes #697 (#735) <Juraci Paixão Kröhling>
@@ -206,7 +283,7 @@ jaeger_query_responses_bucket{operation="find_traces",le="0.005"} 2
 - Track Javascript errors via Google Analytics (fixes [#39](https://github.com/jaegertracing/jaeger-ui/issues/39))
 - Add Google Analytics event tracking for actions in trace view ([#191](https://github.com/jaegertracing/jaeger-ui/issues/191))
 
-##### Fixes
+##### Bug Fixes, Minor Improvements
 
 - Clearly identify traces without a root span (fixes [#190](https://github.com/jaegertracing/jaeger-ui/issues/190))
 - Fix [#166](https://github.com/jaegertracing/jaeger-ui/issues/166) JS error on search page after viewing 404 trace
@@ -227,7 +304,7 @@ jaeger_query_responses_bucket{operation="find_traces",le="0.005"} 2
 - Make agent host port configurable in hotrod (#663) <Pavol Loffay>
 - Add signal handling to standalone (#657) <Pavol Loffay>
 
-##### Fixes
+##### Bug Fixes, Minor Improvements
 
 - Remove the override of GOMAXPROCS (#679) <Cruth kvinc>
 - Use UTC timezone for ES indices (#646) <Pavol Loffay>
@@ -242,7 +319,7 @@ jaeger_query_responses_bucket{operation="find_traces",le="0.005"} 2
   - Fix [#165](https://github.com/jaegertracing/jaeger-ui/issues/165) - Search results are shown without a date
   - Fix [#69](https://github.com/jaegertracing/jaeger-ui/issues/69) - Missing endpoints in jaeger ui dropdown
 
-##### Fixes
+##### Bug Fixes, Minor Improvements
 
 - Fix 2 digit lookback (12h, 24h) parsing (https://github.com/jaegertracing/jaeger-ui/issues/167)
 
@@ -258,7 +335,7 @@ jaeger_query_responses_bucket{operation="find_traces",le="0.005"} 2
 - Add CA certificates to collector/query images (#485)
 - Parse zipkin v2 high trace id (#596)
 
-##### Fixes
+##### Bug Fixes, Minor Improvements
 
 - Skip nil and zero length hits in ElasticSearch storage (#601)
 - Make Cassandra service_name_index inserts idempotent (#587)
@@ -276,7 +353,7 @@ jaeger_query_responses_bucket{operation="find_traces",le="0.005"} 2
 - Add View Option for raw/unadjusted trace (https://github.com/jaegertracing/jaeger-ui/issues/153)
 - Add timezone tooltip to custom lookback form-field (https://github.com/jaegertracing/jaeger-ui/pull/161)
 
-##### Fixes
+##### Bug Fixes, Minor Improvements
 
 - Use consistent icons for logs expanded/collapsed (https://github.com/jaegertracing/jaeger-ui/issues/86)
 - Encode service name in API calls to allow '/' (https://github.com/jaegertracing/jaeger-ui/issues/138)